decimal split

+++ arena.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-mem_arena* arena_create(u64 reserve_size, u64 commit_size) {
+mem_arena* arena_create(size_t reserve_size, size_t commit_size) {
     u32 pagesize = plat_get_pagesize();
 
     reserve_size = ALIGN_UP_POW2(reserve_size, pagesize);
@@ -27,9 +27,9 @@ void arena_destroy(mem_arena* arena) {
     plat_mem_release(arena, arena->reserve_size);
 }
 
-void* arena_push(mem_arena* arena, u64 size, b32 zero) {
-    u64 pos_aligned = ALIGN_UP_POW2(arena->pos, ARENA_ALIGN);
-    u64 new_pos = pos_aligned + size;
+void* arena_push(mem_arena* arena, size_t size, b32 zero) {
+    size_t pos_aligned = ALIGN_UP_POW2(arena->pos, ARENA_ALIGN);
+    size_t new_pos = pos_aligned + size;
 
     if (new_pos > arena->reserve_size) { 
 		fprintf(stderr, "ERROR: out of memory in arena_push\n");
@@ -38,13 +38,13 @@ void* arena_push(mem_arena* arena, u64 size, b32 zero) {
 	}
 
     if (new_pos > arena->commit_pos) {
-        u64 new_commit_pos = new_pos;
+        size_t new_commit_pos = new_pos;
         new_commit_pos += arena->commit_size - 1;
         new_commit_pos -= new_commit_pos % arena->commit_size;
         new_commit_pos = MIN(new_commit_pos, arena->reserve_size);
 
         u8* mem = (u8*)arena + arena->commit_pos;
-        u64 commit_size = new_commit_pos - arena->commit_pos;
+        size_t commit_size = new_commit_pos - arena->commit_pos;
 
         if (!plat_mem_commit(mem, commit_size)) {
 			fprintf(stderr, "ERROR: out of memory in arena_push\n");
@@ -66,13 +66,13 @@ void* arena_push(mem_arena* arena, u64 size, b32 zero) {
     return out;
 }
 
-void arena_pop(mem_arena* arena, u64 size) {
+void arena_pop(mem_arena* arena, size_t size) {
     size = MIN(size, arena->pos - ARENA_BASE_POS);
     arena->pos -= size;
 }
 
-void arena_pop_to(mem_arena* arena, u64 pos) {
-    u64 size = pos < arena->pos ? arena->pos - pos : 0;
+void arena_pop_to(mem_arena* arena, size_t pos) {
+    size_t size = pos < arena->pos ? arena->pos - pos : 0;
     arena_pop(arena, size);
 }
 
@@ -92,20 +92,20 @@ u32 plat_get_pagesize(void) {
     return sysinfo.dwPageSize;
 }
 
-void* plat_mem_reserve(u64 size) {
+void* plat_mem_reserve(size_t size) {
     return VirtualAlloc(NULL, size, MEM_RESERVE, PAGE_READWRITE);
 }
 
-b32 plat_mem_commit(void* ptr, u64 size) {
+b32 plat_mem_commit(void* ptr, size_t size) {
     void* ret = VirtualAlloc(ptr, size, MEM_COMMIT, PAGE_READWRITE);
     return ret != NULL;
 }
 
-b32 plat_mem_decommit(void* ptr, u64 size) {
+b32 plat_mem_decommit(void* ptr, size_t size) {
     return VirtualFree(ptr, size, MEM_DECOMMIT);
 }
 
-b32 plat_mem_release(void* ptr, u64 size) {
+b32 plat_mem_release(void* ptr, size_t size) {
     return VirtualFree(ptr, size, MEM_RELEASE);
 }
 
@@ -122,7 +122,7 @@ u32 plat_get_pagesize(void) {
     return (u32)sysconf(_SC_PAGESIZE);
 }
 
-void* plat_mem_reserve(u64 size) {
+void* plat_mem_reserve(size_t size) {
     void* out = mmap(NULL, size, PROT_NONE, MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
     if (out == MAP_FAILED) {
         return NULL;
@@ -130,19 +130,19 @@ void* plat_mem_reserve(u64 size) {
     return out;
 }
 
-b32 plat_mem_commit(void* ptr, u64 size) {
+b32 plat_mem_commit(void* ptr, size_t size) {
     i32 ret = mprotect(ptr, size, PROT_READ | PROT_WRITE);
     return ret == 0;
 }
 
-b32 plat_mem_decommit(void* ptr, u64 size) {
+b32 plat_mem_decommit(void* ptr, size_t size) {
     i32 ret = mprotect(ptr, size, PROT_NONE);
     if (ret != 0) return false;
     ret = madvise(ptr, size, MADV_DONTNEED);
     return ret == 0;
 }
 
-b32 plat_mem_release(void* ptr, u64 size) {
+b32 plat_mem_release(void* ptr, size_t size) {
     i32 ret = munmap(ptr, size);
     return ret == 0;
 }
